[patch] [docs] feat: Add image example to README
This commit adds a visual example to the README file by including an
image showcasing the output of the GPU simplex noise for 2D generation.
This addition enhances the documentation by providing users with a
clearer understanding of the results they can expect when using
the library.

The image included is `assets/2d_scale50.0_octaves9.jpg`, which
demonstrates the performance improvements of the GPU-based
implementation over traditional CPU methods. This update aims to
improve user engagement and the overall usability of the documentation.